+++ src/misc/validators.ts
@@ -1,7 +1,53 @@
 import {AbstractControl, ValidationErrors, ValidatorFn} from '@angular/forms';
+import {isBlank} from '@jscrpt/common';
 
 import {DateApi} from '../services';
 import {DateTimeValueObject} from '../interfaces/dateTime/datetime.interface';
+import {parseDateTime} from './utils';
+import {DateTimeValueFormat} from './enums';
+import {DateTimeInputOutputValue} from './types';
+
+/**
+ * Date time validator factory function, creates validator for checking validity of datetime
+ * @param dateApi - Date api used for parsing date time
+ * @param valueFormat - Optional required format
+ * @param stringFormat - Optional string format of value
+ */
+export function datetimeValidator<TDate = unknown>(dateApi: DateApi<TDate>,
+                                                   valueFormat: DateTimeValueFormat|undefined|null,
+                                                   stringFormat: string|undefined|null,): ValidatorFn
+{
+    return (control: AbstractControl<DateTimeInputOutputValue<TDate>>): ValidationErrors|null =>
+    {
+        if(isBlank(control.value))
+        {
+            return null;
+        }
+
+        const parsedValue = parseDateTime(control.value, dateApi, valueFormat, stringFormat);
+
+        if(!parsedValue)
+        {
+            return null;
+        }
+
+        if(!Array.isArray(parsedValue))
+        {
+            if(!parsedValue.isValid())
+            {
+                return {
+                    'datetime': true
+                };
+            }
+        }
+        else
+        {
+            //TODO: support for ranges
+        }
+
+        return null;
+    };
+}
 
 /**
  * Validations functions for datetime