+++ src/modules/dateTime/directives/dateTimeInput/dateTimeInput.directive.ts
@@ -1,8 +1,14 @@
-import {Directive, ElementRef, EventEmitter, ExistingProvider, forwardRef, Input, OnDestroy} from '@angular/core';
-import {BindThis} from '@jscrpt/common';
+import {Directive, ElementRef, EventEmitter, ExistingProvider, forwardRef, Inject, Input, OnDestroy} from '@angular/core';
+import {BindThis, isString} from '@jscrpt/common';
 
-import {DateTimeInput, DateTimeValue, FormatProvider} from '../../../../interfaces';
-import {DATE_TIME_INPUT} from '../../../../misc/tokens';
+import {DateTimeInput, FormatProvider} from '../../../../interfaces';
+import {DateTimeValueFormat} from '../../../../misc/enums';
+import {DATE_API, DATE_TIME_INPUT, FORMAT_PROVIDER} from '../../../../misc/tokens';
+import {DateTimeInputOutputValue, DateTimeObjectValue} from '../../../../misc/types';
+import {formatDateTime, parseDateTime} from '../../../../misc/utils';
+import {DateApi} from '../../../../services';
+
+//TODO: range is unimplemented
 
 /**
  * Directive that is used for setting up date time input
@@ -21,6 +27,30 @@ import {DATE_TIME_INPUT} from '../../../../misc/tokens';
 })
 export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, OnDestroy
 {
+    //######################### private fields #########################
+
+    /**
+     * Date time value format which is being worked with in this date time
+     */
+    private _valueFormat: DateTimeValueFormat = DateTimeValueFormat.DateInstance;
+
+    /**
+     * Value of date time
+     */
+    private _value: DateTimeInputOutputValue<TDate>|undefined|null;
+
+    /**
+     * Format of string representation of date
+     */
+    private _format: keyof FormatProvider = 'date';
+
+    //######################### protected properties #########################
+
+    /**
+     * Internal representation of current date time value
+     */
+    protected internalValue: DateTimeObjectValue<TDate>|undefined|null;
+
     //######################### public properties - implementation of DateTimeInput #########################
 
     /**
@@ -38,7 +68,27 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
     /**
      * @inheritdoc
      */
-    public value: string|DateTimeValue<TDate>|TDate|undefined|null;
+    public get value(): DateTimeInputOutputValue<TDate>|undefined|null
+    {
+        return this._value;
+    }
+    public set value(value: DateTimeInputOutputValue<TDate>|undefined|null)
+    {
+        this._value = value;
+
+        //accepts all available formats
+        this.internalValue = parseDateTime(value, this.dateApi, null, this.customFormat);
+        
+        //not range value
+        if(!Array.isArray(this.internalValue))
+        {
+            this.rawValue = this.internalValue?.format(this.customFormat);
+        }
+        else
+        {
+            //TODO: add support for ranges
+        }
+    }
 
     /**
      * @inheritdoc
@@ -70,25 +120,49 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
     //######################### public properties - inputs #########################
 
     /**
-     * Indication whether value for date time input is represented as range from, to
+     * Gets or sets date time value format which is being worked with in this date time
      */
     @Input()
-    public range: boolean = false;
+    public get valueFormat(): DateTimeValueFormat
+    {
+        return this._valueFormat;
+    }
+    public set valueFormat(value: DateTimeValueFormat)
+    {
+        if(isString(value))
+        {
+            this._valueFormat = DateTimeValueFormat[value] as unknown as DateTimeValueFormat;
+
+            return;
+        }
+
+        this._valueFormat = value;
+    }
 
     /**
-     * Indication whether value for date time input is represented as formatted string value, if both 'formatted' and 'range' are set, 'range' takes precedence
+     * Gets or sets format of string representation of date
      */
     @Input()
-    public formatted: boolean = false;
+    public get format(): keyof FormatProvider
+    {
+        return this._format;
+    }
+    public set format(value: keyof FormatProvider)
+    {
+        this._format = value;
+        this.customFormat = this.dateApi.getFormat(this.formatProvider[value]);
+    }
 
     /**
-     * 
+     * Custom format string representation of date
      */
     @Input()
-    public format: keyof FormatProvider|string = 'date';
+    public customFormat: string = this.dateApi.getFormat(this.formatProvider[this._format]);
 
     //######################### constructors #########################
-    constructor(protected element: ElementRef<HTMLInputElement>,)
+    constructor(protected element: ElementRef<HTMLInputElement>,
+                @Inject(DATE_API) protected dateApi: DateApi<TDate>,
+                @Inject(FORMAT_PROVIDER) protected formatProvider: FormatProvider,)
     {
         this.element.nativeElement.addEventListener('input', this.handleInput);
         this.element.nativeElement.addEventListener('focus', this.handleFocus);
@@ -115,6 +189,9 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
     @BindThis
     protected handleInput(): void
     {
+        this.internalValue = parseDateTime(this.rawValue, this.dateApi, this.valueFormat, this.customFormat);
+        this._value = formatDateTime(this.internalValue, this.valueFormat, this.customFormat);
+
         this.valueChange.next();
     }
 
@@ -137,4 +214,11 @@ export class DateTimeInputDirective<TDate = unknown> implements DateTimeInput, O
     {
         this.blur.emit(event);
     }
+
+    //######################### ng language server #########################
+    
+    /**
+     * Custom input type for `valueFormat` input
+     */
+    public static ngAcceptInputType_valueFormat: keyof typeof DateTimeValueFormat|DateTimeValueFormat;
 }