+++ src/modules/dateTime/directives/index.ts
@@ -1,2 +1,3 @@
 export * from './dateTimeControlValueAccessor/dateTimeControlValueAccessor.directive';
 export * from './dateTimeInput/dateTimeInput.directive';
+export * from './dateTimeValidator/dateTimeValidator.directive';